requirement.txt
boto3
requests
pydantic
PyJWT
faiss-cpu
word2number
transformers
datasets
accelerate
